requirements.txt
django~=3.2
djangorestframework~=3.12.4
django-cors-headers
factory-boy
Faker~=8.0.0